requirements.txt
tensorflow==2.16.1
numpy
scipy
matplotlib
pillow
PyQt5
